+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [O.6.0] - unreleased
+## [O.6.0] - 2026-08-20
 
 ### Added
 - LICENSE and NOTICE for Apache 2.0 license
+- `ndjson` parser
 
 ### Changed
 - `now` static method for `Datetime` to support `timezone` parameter.