+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [O.6.2] - 2026-08-21
+
+### Changed
+- no API changes, only bundler fixes.
+
 ## [O.6.1] - 2026-08-21
 
 ### Changed